added MailFolder, moved trash,inbox,sent screens to mailFolderScreen
