+++ appium/tests/screenobjects/all-screens.ts
@@ -1,6 +1,5 @@
 import SplashScreen from './splash.screen';
 import SetupKeyScreen from './setup-key.screen';
-import InboxScreen from './inbox.screen';
 import MenuBarScreen from './menu-bar.screen';
 import EmailScreen from './email.screen';
 import NewMessageScreen from './new-message.screen';
@@ -10,13 +9,11 @@ import ContactScreen from './contacts.screen';
 import ContactPublicKeyScreen from './contact-public-key.screen';
 import PublicKeyScreen from './public-key.screen';
 import AttachmentScreen from './attachment.screen';
-import SentScreen from './sent.screen';
-import TrashScreen from './trash.screen'
+import MailFolderScreen from './mail-folder.screen'
 
 export {
   SplashScreen,
   SetupKeyScreen,
-  InboxScreen,
   MenuBarScreen,
   EmailScreen,
   NewMessageScreen,
@@ -26,6 +23,5 @@ export {
   PublicKeyScreen,
   ContactScreen,
   ContactPublicKeyScreen,
-  SentScreen,
-  TrashScreen
+  MailFolderScreen
 };
