+++ appium/tests/specs/composeEmail/SendEmailToRecipientWithoutPublicKey.spec.ts
@@ -1,7 +1,7 @@
 import {
   SplashScreen,
   SetupKeyScreen,
-  InboxScreen,
+  MailFolderScreen,
   NewMessageScreen
 } from '../../screenobjects/all-screens';
 
@@ -18,9 +18,9 @@ describe('COMPOSE EMAIL: ', () => {
 
     await SplashScreen.login();
     await SetupKeyScreen.setPassPhrase();
-    await InboxScreen.checkInboxScreen();
+    await MailFolderScreen.checkInboxScreen();
 
-    await InboxScreen.clickCreateEmail();
+    await MailFolderScreen.clickCreateEmail();
     await NewMessageScreen.composeEmail(noPublicKeyRecipient, emailSubject, emailText);
     await NewMessageScreen.checkFilledComposeEmailInfo(noPublicKeyRecipient, emailSubject, emailText);
     await NewMessageScreen.clickSendButton();
