+++ appium/tests/specs/inbox/CheckEncryptedEmailAfterRestartApp.spec.ts
@@ -1,7 +1,7 @@
 import {
   SplashScreen,
   SetupKeyScreen,
-  InboxScreen,
+  MailFolderScreen,
   EmailScreen
 } from '../../screenobjects/all-screens';
 
@@ -22,15 +22,15 @@ describe('INBOX: ', () => {
 
     await SplashScreen.login();
     await SetupKeyScreen.setPassPhrase();
-    await InboxScreen.checkInboxScreen();
+    await MailFolderScreen.checkInboxScreen();
 
-    await InboxScreen.clickOnEmailBySubject(emailSubject);
+    await MailFolderScreen.clickOnEmailBySubject(emailSubject);
     await EmailScreen.checkOpenedEmail(senderEmail, emailSubject, emailText);
 
     await driver.terminateApp(bundleId);
     await driver.activateApp(bundleId);
 
-    await InboxScreen.clickOnEmailBySubject(emailSubject);
+    await MailFolderScreen.clickOnEmailBySubject(emailSubject);
 
     //try to see encrypted message with wrong pass phrase
     await EmailScreen.enterPassPhrase(wrongPassPhrase);
@@ -44,7 +44,7 @@ describe('INBOX: ', () => {
 
     //reopen email without pass phrase
     await EmailScreen.clickBackButton();
-    await InboxScreen.clickOnEmailBySubject(emailSubject);
+    await MailFolderScreen.clickOnEmailBySubject(emailSubject);
     await EmailScreen.checkOpenedEmail(senderEmail, emailSubject, emailText);
   });
 });
