+++ appium/tests/specs/settings/CheckSettingsForLoggedUser.spec.ts
@@ -5,7 +5,7 @@ import {
   SettingsScreen,
   KeysScreen,
   PublicKeyScreen,
-  InboxScreen
+  MailFolderScreen
 } from '../../screenobjects/all-screens';
 
 
@@ -15,7 +15,7 @@ describe('SETTINGS: ', () => {
 
     await SplashScreen.login();
     await SetupKeyScreen.setPassPhrase();
-    await InboxScreen.checkInboxScreen();
+    await MailFolderScreen.checkInboxScreen();
 
     await MenuBarScreen.clickMenuIcon();
     await MenuBarScreen.checkUserEmail();
